Implement Dapr event publishing on counter updates
Added a line to publish a Dapr event after updating the counter in the Counter.razor component. Commented out the initial Dapr store and pubsub setup code in the AppHost Program.cs file, as the setup there was likely replaced by the new Dapr.AspNetCore package and code added to the CountService. Also revised CountService Program.cs to define a subscriber for the "counter" topic.