Implement Fluent UI in OpenAI webpages and add Labels page
The commit includes the replacement of Bootstrap with Microsoft Fluent UI for Blazor, an improved and consistent user interface. It restructures the Home page and NavMenu to adapt Fluent UI components. Also, a new page called 'Labels' has been added to the set. It introduces changes in few .css files for design enhancements. Additionally, the Summarize.razor page is deleted.